Disable clear canvas for WebGL between frames
Since we are using a background and always drawing the background there
is no need for WebGL to clear the canvas before the next frame. The
background will just draw on top of all previous draws anyways.
It's not the biggest performance boost but it's such an easy fix so, why
not.

Source of finding: https://phaser.io/tutorials/advanced-rendering-tutorial/part5